+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/CgibinAccount/CgibinAccountModifyHeadImageRequest.cs
@@ -17,27 +17,27 @@
         /// </summary>
         [Newtonsoft.Json.JsonProperty("x1")]
         [System.Text.Json.Serialization.JsonPropertyName("x1")]
-        public double X1 { get; set; }
+        public decimal X1 { get; set; }
 
         /// <summary>
         /// 获取或设置裁剪框左上角 Y 坐标（范围：0～1）。
         /// </summary>
         [Newtonsoft.Json.JsonProperty("y1")]
         [System.Text.Json.Serialization.JsonPropertyName("y1")]
-        public double Y1 { get; set; }
+        public decimal Y1 { get; set; }
 
         /// <summary>
         /// 获取或设置裁剪框右下角 X 坐标（范围：0～1）。
         /// </summary>
         [Newtonsoft.Json.JsonProperty("x2")]
         [System.Text.Json.Serialization.JsonPropertyName("x2")]
-        public double X2 { get; set; }
+        public decimal X2 { get; set; }
 
         /// <summary>
         /// 获取或设置裁剪框右下角 Y 坐标（范围：0～1）。
         /// </summary>
         [Newtonsoft.Json.JsonProperty("y2")]
         [System.Text.Json.Serialization.JsonPropertyName("y2")]
-        public double Y2 { get; set; }
+        public decimal Y2 { get; set; }
     }
 }
