+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/Wxa/WxaStore/WxaGetDistrictResponse.cs
@@ -18,14 +18,14 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("lng")]
                         [System.Text.Json.Serialization.JsonPropertyName("lng")]
-                        public double Longitude { get; set; }
+                        public decimal Longitude { get; set; }
 
                         /// <summary>
                         /// 获取或设置纬度。
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("lat")]
                         [System.Text.Json.Serialization.JsonPropertyName("lat")]
-                        public double Latitude { get; set; }
+                        public decimal Latitude { get; set; }
                     }
                 }
